opcualogger/pyproject.toml
[project]
name = "opcualogger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncua>=1.1.8",
    "pandas>=3.0.1",
    "pyinstaller>=6.19.0",
    "python-dotenv>=1.2.1",
    "tool>=0.8.0",
]
